https: workingish token requests
With a bit of craftsmanship we can now do token requests based
on the authorization code and handle the response. There's a
few sharp edges here and there, but we should now be able
to parse the response.

The token response looks simple enough that it's actually tempting
to parse it by hand, instead of dropping in a json parsing library.

I suppose we'll soon need a place *cough* sqlite3 *cough* to store
these so we can skip inventing in-memory data structures for
"sessions".