https: support Transfer-Encoding: chunked
It seems youtube wants to send the heavy lists this way, no wonder
parsing proved... tricky.

While at it, a memory leak was fixed in the header parsing code,
where we'd keep reading lines and losing the one we just read.

On the flip side, there's now probably a few more.