feed: A simple interface to expat for parsing video lists
Provides pretty output directly to an event buffer from a
response body.

It's not the most robust thing in the world, most likely leaks
like a sieve and probably goes belly up as soon as google does
something with their kit.

But it sort of works and there's css embedded in it too, for
extra prettiness.

Plugging navigation into this should be a breeze.