list: Hook the feed parser into list_handle()
Oh woo, managed to come up with an interface to the parser
that only requires flipping argument order in the callback.

Yay me.

There's a lot of duplicates in the list too, wonder what to
do with those.

Perhaps we should grab videoid and do uniqueness checks based on the
video id or something, updating F_PUBLISHED/F_UPDATED accordingly. But
this would mean not being able to spew entries at the browser the minute
they appear.

Not that they appear immediately anyway, as we're kind of doing this in
the thread trying to handle the request...