test: test things locally instead of banging head against google
Throw in CUnit and a simple test that verifies we actually render
navigation links.

A note for kids doing unit tests at home, this is not how you do
it. We actually consume the whole feed, then parse the resulting
html and try to find the previous/next links.

And since expat is more strict than browsers, we get to fix the
html we push out :|