store: propagate error from session_ensure() up to caller
Propagate ENOMEM from session_ensure() upwards and do proper cleanup
when the store is full.

The caller is pretty much out of options at this stage:

    /--->  store full
   |           -> no session
   |               -> no cookie
   |                   -> error to browser
   |                       -> retry
   |
   |                            |
    \--------------------------/

No idea if we should try to reuse older session entries, as we have the
necessary bookkeeping. But then again, that's local DoS :)

One more bug and these hcreate() things are out the door. They're not
worth it.