https: let BIO_free() release the underlying transport
For some reason bufferevent_openssl_socket_new() pokes the bio
backing the ssl object we pass in, doing BIO_set_close(bio, 0).

This results in leaked memory for every request; The bio itself is
freed, the ssl object is freed, but as openssl has been instructed
to not close the underlying stream, it's lost.

Fix it by BIO_set_close(bio, 1) right before we do bufferevent_free()
which frees the bio.