https: detect end of request in the read callback
Introduce a new read state, READ_DONE, for situations where we
detect we've read the whole response.

  a) For "Transfer-Encoding: chunked" this is indicated by a
     0-length chunk

  b) For "normal" responses we check if Content-Length is set
     and we've consumed Content-Length bytes.

This prepares us for the situation where "we're done" might not
be indicated by the remote end (or us) closing the connection ;)

Also while we're in READ_BODY, loop until we've exhausted the input
buffer. We don't have level-triggering so we'd better deal with all
the data before returning from the read callback. We may have
problems with this still but at least we're not hitting them.