https: add a new READ_NONE state, (cb_read() never called)
To have a sane heuristic for when we should restart the request
we'll be using this.

We can't use any write counters, as it seems we manage to write()
the request even when it's a closed connection, we'll just get an
error soon after.

This is definitely something that wants refining later.