https: copy request body instead of losing it
In order to do request restarts transparently without burdening
https_request() callers we cannot lost the request body.

We used to use evbuffer_add_buffer() for copying the request body
into output. Doing this drains the original buffer, which is not
acceptable.

So store request body in struct request_ctx as a proper data blob
and do evbuffer_add_reference() on it when we construct the output
buffer. We don't need to specify cleanup functions as we only get
rid of the struct request_ctx when we know nothing's touching the
bufferevent.