keepalive: reconnect keepalive connections that have been closed
The heuristics for "are we disconnected" probably aren't perfect,
but reconnecting a timed out connection and replacing the stashed
connection with the new fresh one looks to be working.

All of this takes place behind the https_request() curtains, so the
caller will never know what goes on.

I suppose we'll need a limit for retries anyway, just to be sure.