Implement the no-keepalive debugging option
Passing -n will cause https keepalive to be disabled. We do this
by passing Connection: close on requests and just discarding
connections at conn_stash_put_bev() instead of stashing them away.

This has the obvious effect of conn_stash_get_bev() never finding a
stashed connection and always returning a fresh one.

And lo and behold, there's a nasty wart in https.c:cb_event() again
to deal with end-of-response :(

Also a fair amount of noise in the diff as we pass the argument all
the way down from main via https to conn_stash.