https: consider non-200 status from YouTube as an error
When trying to authorize a Google account that's _not_ linked to a
YouTube account, the API tells us "401 NoLinkedYoutubeAccount".

Consider anything non-200 as a "hard" error and propagate that to
the browser like we would the other ones too.

Of course we could also pass the response status to the regular
->read_data() callbacks and the like, but that would be.. more
involved and this works just as well here.